Basic tests done
